refactor(script): split remaining misc buildins into buildin_misc.c
Issue #61 twelfth step: move leftover helpers out of script.c while
keeping the script engine core; share sql accessors via script_internal.h
and replace getmapxy/sqlquery str_buf access with get_str().
