+++ vcproj/map-server.vcxproj.filters
@@ -139,6 +139,9 @@
     <ClCompile Include="..\src\map\buildin_md.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_misc.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
       <Filter>map_sql</Filter>
     </ClCompile>
